Fix: output of codes
