+++ langchain/utilities/normal_captures.py
@@ -4,10 +4,9 @@ class Piece:
         self.value = value
         self.is_dama = is_dama
         self.index = index
-        self.name = f"{color}, {value}"
 
     def __repr__(self):
-        return f"Piece('{self.color}', {self.value}, {self.is_dama})"
+        return f"Piece('{self.color}', {self.value}, is_dama={self.is_dama})"
 
     def __eq__(self, other):
         if not isinstance(other, Piece):
@@ -18,34 +17,84 @@ class Piece:
         return hash(self.name)
 
 
-# Define the board state
-board_state = [[Piece('r', -112, is_dama=False), '*'], 'X', [None, '/'], 'X', [None, '-'], 'X', [None, '+'], 'X', 'X',
-               [Piece('b', 0, is_dama=False), '/'], 'X', [None, '*'], 'X', [None, '+'], 'X', [None, '-'], [None, '-'],
-               'X', [None, '+'], 'X', [None, '*'], 'X', [Piece('r', -9, is_dama=False), '/'], 'X', 'X',
-               [None, '+'], 'X', [Piece('b', -11, is_dama=False), '-'], 'X', [None, '/'], 'X', [None, '*'],
-               [Piece('b', 0, is_dama=False), '*'], 'X', [Piece('r', -5, is_dama=True), '/'], 'X', [None, '-'],
-               'X', [None, '+'], 'X', 'X', [None, '/'], 'X', [Piece('b', -5, is_dama=True), '*'], 'X',
-               [None, '+'], 'X', [None, '-'], [None, '-'], 'X', [None, '+'], 'X', [None, '*'],
-               'X', [Piece('b', 6, is_dama=False), '/'], 'X', 'X', [None, '+'], 'X', [None, '-'],
-               'X', [None, '/'], 'X', [Piece('r', 6, is_dama=False), '*']]
+board_state = [
+    [Piece('r', -112, is_dama=False), '*'],
+    'X',
+    [None, '/'],
+    'X',
+    [None, '-'],
+    'X',
+    [None, '+'],
+    'X', 'X',
+    [Piece('b', 0, is_dama=False), '/'],
+    'X',
+    [None, '*'],
+    'X',
+    [None, '+'],
+    'X',
+    [None, '-'],
+    [None, '-'],
+    'X',
+    [None, '+'],
+    'X',
+    [None, '*'],
+    'X',
+    [Piece('r', -9, is_dama=False), '/'],
+    'X', 'X',
+    [None, '+'],
+    'X',
+    [Piece('b', -11, is_dama=False), '-'],
+    'X',
+    [None, '/'],
+    'X',
+    [None, '*'],
+    [Piece('b', 0, is_dama=False), '*'],
+    'X',
+    [Piece('r', -5, is_dama=True), '/'],
+    'X',
+    [None, '-'],
+    'X',
+    [None, '+'],
+    'X', 'X',
+    [None, '/'],
+    'X',
+    [Piece('b', -5, is_dama=True), '*'],
+    'X',
+    [None, '+'],
+    'X',
+    [None, '-'],
+    [None, '-'],
+    'X',
+    [None, '+'],
+    'X',
+    [None, '*'],
+    'X',
+    [Piece('b', 6, is_dama=False), '/'],
+    'X', 'X',
+    [None, '+'],
+    'X',
+    [None, '-'],
+    'X',
+    [None, '/'],
+    'X',
+    [Piece('r', 6, is_dama=False), '*']
+]
 
 
 def func5(board_state):
     valid_moves = {}
-    
-    for i, element in enumerate(board_state):
-        if isinstance(element, str) and element == 'X':
+    for i, elem in enumerate(board_state):
+        if isinstance(elem, str) and elem == 'X':
             continue
-        elif isinstance(element, list):
-            first_element = element[0]
-            if first_element is None:
+        elif isinstance(elem, list):
+            first_elem = elem[0]
+            if first_elem is None:
                 continue
-            elif isinstance(first_element, Piece) and first_element.color == 'r':
-                piece = first_element
+            if isinstance(first_elem, Piece) and first_elem.color == 'r':
+                piece = first_elem
                 piece.index = i
                 key, value = func6(piece)
                 valid_moves.update({key: value})
-    
     return valid_moves
 
 
@@ -53,21 +102,26 @@ def func6(piece):
     capmoves = []
     dia = [7, 9, -7, -9]
     src_ind = piece.index
-    
+
     for d in dia:
         temp_ind = src_ind + d
         if not 0 <= temp_ind < len(board_state):
             continue
-        
         if isinstance(board_state[temp_ind], list) and board_state[temp_ind][0] is not None:
-            dest_ind = temp_ind
-            if board_state[dest_ind][0].color == 'b':
-                capmoves.append(dest_ind)
-    
+            first_piece = board_state[temp_ind][0]
+            if first_piece.color == 'b':
+                dest_ind = temp_ind + d
+                if 0 <= dest_ind < len(board_state):
+                    nested_list = board_state[dest_ind]
+                    if isinstance(nested_list, list) and nested_list[0] is None:
+                        capmoves.append(dest_ind)
     return piece.index, capmoves
 
 
-# Call func5 with the board state
-result = func5(board_state)
+def func7(valid_moves):
+    # This function should update the valid moves dictionary
+    pass  # For now, it's left empty as there are no instructions for what to do with this variable
 
-print(result)
+
+# Call func5 and print the result
+print(func5(board_state))
