+++ langchain/utilities/normal_moves.py
@@ -1,12 +1,13 @@
 class Piece:
-    def __init__(self, color, value, is_dama=False):
+    def __init__(self, color, value, is_dama=False, index=0):
         self.color = color
         self.value = value
         self.is_dama = is_dama
-        self.index = 0
+        self.index = index
+        self.name = f"{color}, {value}"
 
     def __repr__(self):
-        return f"Piece('{self.color}', {self.value}, is_dama={self.is_dama})"
+        return f"Piece('{self.color}', {self.value}, {self.is_dama})"
 
     def __eq__(self, other):
         if not isinstance(other, Piece):
@@ -16,46 +17,99 @@ class Piece:
     def __hash__(self):
         return hash(self.name)
 
+
 def func1(board_state):
     valid_moves = {}
-    
     for i, element in enumerate(board_state):
-        if type(element) is list:
-            piece = None
-            if element[0] is not None and isinstance(element[0], Piece):
-                piece = element[0]
-                if piece.color == 'r':
-                    piece.index = i
-            
-            if piece is not None:
-                func2(piece, board_state)
-        
-        elif type(element) is str:
+        if element == 'X':
             continue
-        
-        else:
-            raise ValueError("Invalid element in the board state")
-    
+        elif isinstance(element, list):
+            first_element = element[0]
+            if first_element is None:
+                continue
+            elif isinstance(first_element, Piece) and first_element.color == "r":
+                piece = first_element
+                piece.index = i
+            key, value = func2(piece)
+            valid_moves.update({key: value})
     return valid_moves
 
-def func2(piece, board_state):
+
+def func2(piece):
     moves = []
-    
-    for destination_index in [piece.index + 7, piece.index + 9]:
-        if destination_index < len(board_state) and isinstance(board_state[destination_index], list):
-            destination_square = board_state[destination_index][0]
-            
-            if destination_square is None:
-                moves.append(destination_index)
-            
-            elif isinstance(destination_square, Piece) or destination_square == 'X':
-                continue
-            
-            else:
-                moves.append(destination_index)
-    
+    for move in [piece.index + 7, piece.index + 9]:
+        if 0 <= move < len(board_state):
+            destination_square = board_state[move]
+            if isinstance(destination_square, list) and destination_square[0] is None:
+                moves.append(move)
     return (piece.index, moves)
 
-board_state = [[Piece('r', 2, is_dama=False), '*'], 'X', [Piece('r', -5, is_dama=False), '/'], 'X', [Piece('r', 8, is_dama=False), '-'], 'X', [Piece('r', -11, is_dama=False), '+'], 'X', 'X', [Piece('r', -7, is_dama=False), '/'], 'X', [Piece('r', 10, is_dama=False), '*'], 'X', [Piece('r', -3, is_dama=False), '+'], 'X', [Piece('r', 0, is_dama=False), '-'], [Piece('r', 4, is_dama=False), '-'], 'X', [Piece('r', -1, is_dama=False), '+'], 'X', [Piece('r', 6, is_dama=False), '*'], 'X', [Piece('r', -9, is_dama=False), '/'], 'X', 'X', [None, '+'], 'X', [None, '-'], 'X', [None, '/'], 'X', [None, '*'], [None, '*'], 'X', [None, '/'], 'X', [None, '-'], 'X', [None, '+'], 'X', 'X', [Piece('b', -9, is_dama=False), '/'], 'X', [Piece('b', 6, is_dama=False), '*'], 'X', [Piece('b', -1, is_dama=False), '+'], 'X', [Piece('b', 4, is_dama=False), '-'], [Piece('b', 0, is_dama=False), '-'], 'X', [Piece('b', -3, is_dama=False), '+'], 'X', [Piece('b', 10, is_dama=False), '*'], 'X', [Piece('b', -7, is_dama=False), '/'], 'X', 'X', [Piece('b', -11, is_dama=False), '+'], 'X', [Piece('b', 8, is_dama=False), '-'], 'X', [Piece('b', -5, is_dama=False), '/'], 'X', [Piece('b', 2, is_dama=False), '*']]
-    
-    return func1(board_state)
+
+board_state = [
+    [Piece('r', 2, is_dama=False), '*'],
+    'X',
+    [Piece('r', -5, is_dama=False), '/'],
+    'X',
+    [Piece('r', 8, is_dama=False), '-'],
+    'X',
+    [Piece('r', -11, is_dama=False), '+'],
+    'X',
+    'X',
+    [Piece('r', -7, is_dama=False), '/'],
+    'X',
+    [Piece('r', 10, is_dama=False), '*'],
+    'X',
+    [Piece('r', -3, is_dama=False), '+'],
+    'X',
+    [Piece('r', 0, is_dama=False), '-'],
+    [Piece('r', 4, is_dama=False), '-'],
+    'X',
+    [Piece('r', -1, is_dama=False), '+'],
+    'X',
+    [Piece('r', 6, is_dama=False), '*'],
+    'X',
+    [Piece('r', -9, is_dama=False), '/'],
+    'X',
+    'X',
+    [None, '+'],
+    'X',
+    [None, '-'],
+    'X',
+    [None, '/'],
+    'X',
+    [None, '*'],
+    [None, '*'],
+    'X',
+    [None, '/'],
+    'X',
+    [None, '-'],
+    'X',
+    [None, '+'],
+    'X',
+    'X',
+    [Piece('b', -9, is_dama=False), '/'],
+    'X',
+    [Piece('b', 6, is_dama=False), '*'],
+    'X',
+    [Piece('b', -1, is_dama=False), '+'],
+    'X',
+    [Piece('b', 4, is_dama=False), '-'],
+    [Piece('b', 0, is_dama=False), '-'],
+    'X',
+    [Piece('b', -3, is_dama=False), '+'],
+    'X',
+    [Piece('b', 10, is_dama=False), '*'],
+    'X',
+    [Piece('b', -7, is_dama=False), '/'],
+    'X',
+    'X',
+    [Piece('b', -11, is_dama=False), '+'],
+    'X',
+    [Piece('b', 8, is_dama=False), '-'],
+    'X',
+    [Piece('b', -5, is_dama=False), '/'],
+    'X',
+    [Piece('b', 2, is_dama=False), '*']
+]
+
+print(func1(board_state))
